requirements.txt
torch==2.2.1+cu121
torchvision==0.17.1+cu121
pillow==9.4.0
facenet-pytorch==2.5.3
pre-commit==3.7.1
dvc==3.51.1
hydra-core==1.3.2
dvc-gdrive==3.0.1
pytorch-lightning==2.2.5
